hoàn thiện chức năng đánh giá

+++ database/migrations/2020_01_03_081111_create_rates_table.php
@@ -18,7 +18,7 @@ class CreateRatesTable extends Migration
             $table->bigInteger('product_id');
             $table->bigInteger('customer_id');
             $table->string('comment');
-            $table->float('rate');
+            $table->integer('rate');
             $table->timestamps();
         });
     }
